MergeTwoSortedList
This is Problem No. 21 on LeetCode and in this problem we are given the heads of two sorted linked lists list1 and list2 and we need to merge the two lists in a one sorted list and at last we need to return the head of the merged linked list.